WIP: create user needs testing
