+++ api_dev/crud/crud.py
@@ -1,7 +1,7 @@
 from sqlalchemy.orm import Session
-
 import models
 import schemas
+import hashing
 
 
 def get_user_by_id(db: Session, user_id: int):
@@ -17,8 +17,8 @@ def get_users(db: Session, skip: int = 0, limit: int = 100):
 
 
 def create_user(db: Session, user: schemas.UserCreate):
-    fake_hashed_password = user.password + "notreallyhashed"
-    db_user = models.User(email=user.email, hashed_password=fake_hashed_password)
+    hashed_password = hashing.get_password_hash(user.password)
+    db_user = models.User(email=user.email, hashed_password=hashed_password)
     db.add(db_user)
     db.commit()
     db.refresh(db_user)
